# backend/ffg/openapi.yml
openapi: 3.0.3
info:
  title: FFG Mentorship Platform API
  description: Complete OpenAPI contract specification for the FFG Corporate Mentorship Platform backend, documenting all model CRUD viewsets, custom workflow actions, and analytical dashboards.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Local Development Server
...
paths:
...
  /api/users/:
    get:
      tags:
        - Users
      summary: List all users
      responses:
        '200':
          description: A list of platform operational user records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create a user record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created.
...
  /api/applications/:
    get:
      tags:
        - Applications
      summary: List applications
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
    post:
      tags:
        - Applications
      summary: Create an application entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
      responses:
        '201':
          description: Application recorded.
...
components:
  schemas:
...
    User:
      type: object
      required:
        - username
        - email
        - role
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        role:
          type: string
          enum: [manager, mentor, mentee]
          example: mentee
...
    Application:
      type: object
      required:
        - user
        - cohort
        - form_responses
      properties:
        id:
          type: integer
          readOnly: true
        user:
          type: integer
        cohort:
          type: integer
        status:
          type: string
          enum: [draft, submitted, approved, rejected]
          default: draft
        form_responses:
          type: object
          description: Multi-field survey response value collection mappings JSON payload.
        created_at:
          type: string
          format: date-time
          readOnly: true
...